breaking down the functions
so as requested in my last tutor call i needed to take a look at the function i had made which had become to large and made it difficult to identify where the issue was. i have also included console logs on the start game function and the handleEnd one as well. still getting errors and doesnt seem like the questions are displaying right so may need to book another call